refactor(combo): split transition state and slot binding into core modules
core/ComboTransition consolidates foreground/handoff state - g_PendingMMFileNum,
g_pendingOOTReturn, g_mmReturnKind, g_MmSaveInMemorySlot - which had six declaration sites for
one concept, plus SetForegroundGame, the erase/copy notify wrappers and the return callbacks.

core/ComboSlotBind takes the new-file and load-file callbacks; its forward declaration of
Combo_ResumeMMIfLastSavedThere is now the transition header.

ComboShip.cpp is main() and the DLL handle.
