+++ combo/ComboShip.cpp
@@ -48,216 +48,13 @@
 #include "core/ComboReload.h"
 #include "core/ComboGoal.h"
 #include "core/ComboHintReveal.h"
+#include "core/ComboSlotBind.h"
+#include "core/ComboTransition.h"
 
-// OOT slot whose MM save is live in MM's dormant memory (-1 = none). Guards Combo_OnOOTSaveLoad
-// against reloading stale disk state over MM's in-memory progress after round trips.
-int g_MmSaveInMemorySlot = -1;
-static bool g_pendingOOTReturn = false;
 static DllHandle comboUIModule = nullptr;
 
-// Single place the foreground game changes, so every transition point notifies comboui consistently.
-static void Combo_SetForegroundGame(int game) {
-    // #173: MM only accrues play time while it is foreground. OOT owns the foreground at startup.
-    static int sPrevGame = ComboRando::GAME_OOT;
-    if (sPrevGame == ComboRando::GAME_MM && MM_ComboPausePlaytime)
-        MM_ComboPausePlaytime();
-    if (game == ComboRando::GAME_MM && MM_ComboResumePlaytime)
-        MM_ComboResumePlaytime();
-    sPrevGame = game;
-    if (ComboUI_OnForegroundGame)
-        ComboUI_OnForegroundGame(game);
-}
 
-// Erase/copy the slot's container, then tell the DLLs. ComboContainer owns the storage but cannot make
-// these calls itself (it never sees the export table), which is what keeps them outside the lock.
-static void EraseComboContainer(int slot) {
-    ComboEraseSlotStorage(slot);
-    // #182: MM caches which slot's owl save its gSaveContext came from; that section is gone.
-    if (MM_InvalidateOwlBlobSlot)
-        MM_InvalidateOwlBlobSlot();
-    // #164: the Hint Tracker would otherwise keep showing the deleted slot's hints on file-select.
-    if (ComboUI_SetHintTrackerData)
-        ComboUI_SetHintTrackerData(-1, "", "");
-}
 
-// OOT file-select "copy file": the .combosav has no per-game file to copy, so the launcher owns it.
-static void Combo_CopyContainer(int from, int to) {
-    ComboCopySlotStorage(from, to);
-    // #182: the destination's owl save came from the donor, so MM's descent cache is wrong.
-    if (MM_InvalidateOwlBlobSlot)
-        MM_InvalidateOwlBlobSlot();
-}
-
-// ComboShip (issue #1): erasing a slot from either game's file-select wipes BOTH saves — each game
-// fires its Set*-registered callback with the slot, the launcher routes it to the other game's
-// save-only delete export (never re-entering a menu erase path). See docs/deviations/boot-shutdown.md.
-
-// Registered into each game; invoked when that game erases a slot. Routes the (0-based) slot to the
-// OTHER game's delete export, then removes the merged container. The launcher does no index math —
-// MM's 1-based JSON naming is handled inside MM_DeleteSaveFile.
-static void DeleteForeignSaveFromOOT(int slot) {
-    if (MM_DeleteSaveFile)
-        MM_DeleteSaveFile(slot);
-    EraseComboContainer(slot); // remove the slot's merged container (baked rando + both saves)
-}
-static void DeleteForeignSaveFromMM(int slot) {
-    if (SOH_DeleteSaveFile)
-        SOH_DeleteSaveFile(slot);
-    EraseComboContainer(slot);
-}
-
-static void Combo_OnOOTSaveInit(int fileNum) {
-    // A new file starts in OOT. Explicit because not every delete path clears the container
-    // (DeleteFileOnDeath calls DeleteZeldaFile directly), so a stale MM could otherwise survive here.
-    Combo_SetLastGame(fileNum, ComboRando::GAME_OOT);
-    // ComboShip (#165/#164): a fresh file starts with an empty personal note (written, never erased —
-    // an absent key is the "never migrated" sentinel) and must not inherit the hint read state.
-    ComboResetSlotForNewFile(fileNum);
-    // ComboShip: bind the pending consolidated seed to this slot — bake it into the container's
-    // combo.rando (self-contained), then push it into both DLLs so foreign data is live immediately.
-    nlohmann::json seed;
-    if (!g_ConsolidatedJson.empty()) {
-        try {
-            seed = nlohmann::json::parse(g_ConsolidatedJson);
-        } catch (...) {}
-        ComboBakeSeed(fileNum, seed);
-        LoadComboCompletion(fileNum); // refresh the in-memory latch + push this seed's goal into both DLLs
-        if (SOH_LoadComboRando)
-            SOH_LoadComboRando(g_ConsolidatedJson.c_str());
-        if (MM_LoadComboRando)
-            MM_LoadComboRando(g_ConsolidatedJson.c_str());
-        std::cout << "[ComboShip] baked consolidated seed into container for slot " << fileNum << std::endl;
-        // ComboShip (#135): an MM-start seed routes the fresh file into MM; the launcher handoff then
-        // spawns it in South Clock Town (the same place the portal arrives at).
-        if (seed.value("startingGame", std::string("OOT")) == "MM") {
-            Combo_SetLastGame(fileNum, ComboRando::GAME_MM);
-            std::cout << "[ComboShip] slot " << fileNum << " starts in Majora's Mask" << std::endl;
-        }
-    }
-    // The new-save callback runs on OOT's thread with the entered file name current — carry it into
-    // the matching MM save so both files show the player's name.
-    unsigned char playerName[8] = { 0x3E, 0x3E, 0x3E, 0x3E, 0x3E, 0x3E, 0x3E, 0x3E }; // 0x3E = N64 blank glyph
-    if (SOH_GetCurrentPlayerName)
-        SOH_GetCurrentPlayerName(playerName);
-    // Re-derived every creation, never cached — a consumed cache left later files with a vanilla MM
-    // save, silently disabling every IS_RANDO behavior. See docs/deviations/rando.md.
-    std::string mmPlacements;
-    if (!seed.is_null())
-        mmPlacements = ComboRando::ApplyPayloadFromConsolidated(seed, ComboRando::GAME_MM).dump();
-    if (MM_InitRandoSaveFile && !seed.is_null()) {
-        std::cout << "[ComboShip] Creating RANDO MM save for OOT slot " << fileNum << std::endl;
-        // Re-assert prices from the seed being applied — a failed re-generation after a reload leaves
-        // the MM DLL's captured price map holding the failed seed's rolls, not this spoiler's.
-        if (MM_SetCheckPrices)
-            MM_SetCheckPrices(
-                seed.value("mm", nlohmann::json::object()).value("prices", nlohmann::json::object()).dump().c_str());
-        // A reloaded seed's MM settings only get written here (MM_InitRandoSaveFile is where MM reads
-        // them) — never at reload time, so they can't leak into comboship.json before a slot is bound.
-        if (!g_PendingMMSettingsJson.empty() && MM_RestoreRandoSettings)
-            MM_RestoreRandoSettings(g_PendingMMSettingsJson.c_str());
-        if (MM_InitRandoSaveFile(fileNum, mmPlacements.c_str(), playerName) != 0) {
-            std::cerr << "[ComboShip] ERROR: MM rando save creation FAILED for slot " << fileNum
-                      << " — this slot's MM save has no placements. Re-create it." << std::endl;
-        } else if (ComboUI_SyncRandomizedCosmetics) {
-            // #169: MM has just rolled its own cosmetics (generation hook inside the call above); with
-            // sync on, hand it OOT's colors instead so the shared elements match. Roll OOT first: the
-            // latch skipped it at load time if the options were off, so enabling them mid-seed still
-            // syncs this seed's fresh colors rather than whatever was persisted.
-            if (ComboUI_CosmeticsSyncGateEnabled && ComboUI_CosmeticsSyncGateEnabled())
-                Combo_FireGenRollHooksOnce(seed.value("masterSeed", 0u));
-            ComboUI_SyncRandomizedCosmetics();
-        }
-        // Silent auto-load: the save now has the seed's settings baked in — return the CVars to the
-        // user's config. An explicit drop leaves the seed's settings as the new persisted baseline.
-        if (g_ComboReloadRestoreUserMM && MM_RestoreRandoSettings && !g_UserMMSettingsSnapshot.empty())
-            MM_RestoreRandoSettings(g_UserMMSettingsSnapshot.c_str());
-        g_PendingMMSettingsJson.clear();
-        g_UserMMSettingsSnapshot.clear();
-        g_ComboReloadRestoreUserMM = false;
-    } else {
-        // No seed bound to this slot. ComboShip has no vanilla mode, so there is no valid save to
-        // create here — fail loudly instead of writing one that looks fine and misbehaves later.
-        std::cerr << "[ComboShip] ERROR: no consolidated seed for slot " << fileNum
-                  << " — cannot create an MM rando save. Generate or load a seed first." << std::endl;
-    }
-    // ComboShip (#164): the slot's hints are baked and its read state freshly erased — hand both to
-    // comboui's Hint Tracker.
-    Combo_PushHintTrackerData(fileNum);
-    // The creation path builds the save in MM's live gSaveContext.
-    g_MmSaveInMemorySlot = fileNum;
-}
-
-static void Combo_ResumeMMIfLastSavedThere(int fileNum);
-
-// ComboShip: OOT loaded a save (file select / warp). Bring the matching MM save into MM's dormant
-// memory so the combo tracker peek shows real MM items before MM is visited. Skipped when that
-// slot's MM save is already live in memory — reloading from disk would clobber newer progress.
-static void Combo_OnOOTSaveLoad(int fileNum) {
-    LoadComboCompletion(fileNum); // refresh both-bosses-beaten flags for this slot
-    // Push the slot's baked combo rando (foreign map + cross-hints) into both DLLs, once per load.
-    {
-        const std::string rando = ComboReadBakedRando(fileNum);
-        if (!rando.empty()) {
-            if (SOH_LoadComboRando)
-                SOH_LoadComboRando(rando.c_str());
-            if (MM_LoadComboRando)
-                MM_LoadComboRando(rando.c_str());
-        }
-    }
-    Combo_PushHintTrackerData(fileNum); // #164: this slot's hints + persisted read state
-    if (!MM_LoadSaveForCombo || g_MmSaveInMemorySlot == fileNum) {
-        Combo_OnTriforceProgress(0, fileNum);
-        Combo_ResumeMMIfLastSavedThere(fileNum);
-        return;
-    }
-    std::cout << "[ComboShip] Loading MM save for OOT slot " << fileNum << " (tracker peek)" << std::endl;
-    // Read-only peek: on failure nothing was loaded, so the slot must NOT be marked resident — stale
-    // dormant memory would otherwise pose as this slot's save (and a dormant write would persist it).
-    if (MM_LoadSaveForCombo(fileNum) == 0) {
-        g_MmSaveInMemorySlot = fileNum;
-    }
-    // Both counters are now live: catch a goal crossed while the game wasn't running (e.g. a teammate's
-    // pieces applied to a dormant save). Latched, so it can't roll credits twice.
-    Combo_OnTriforceProgress(0, fileNum);
-    Combo_ResumeMMIfLastSavedThere(fileNum);
-}
-
-// ComboShip (#89): resume MM instead of starting OOT when the slot was last played in MM. The
-// file-select gate is load-bearing — OnLoadGame also fires on the MM->OOT return and from in-game
-// reloads, where this would bounce the player back into MM forever.
-static void Combo_ResumeMMIfLastSavedThere(int fileNum) {
-    if (!SOH_ParkForComboMMResume || !MM_RunGame || !SOH_IsOnFileSelect || !SOH_IsOnFileSelect()) {
-        return;
-    }
-    if (fileNum < 0 || fileNum > 2) {
-        return; // debug select (0xFF) / Boss Rush (0xFE) share FileChoose_LoadGame
-    }
-    if (Combo_GetLastGame(fileNum) != ComboRando::GAME_MM) {
-        return;
-    }
-    std::cout << "[ComboShip] Slot " << fileNum << " was last saved in MM — resuming MM" << std::endl;
-    if (MM_SetComboEntryIsResume)
-        MM_SetComboEntryIsResume(1); // a real save load: honors MM's Remember Save Location
-    g_PendingMMFileNum = fileNum;
-    SOH_ParkForComboMMResume(); // drops out of OOT's game loop; the launcher then enters MM
-}
-
-static void Combo_OnOOTSceneSwitch(int fileNum) {
-    std::cout << "[ComboShip] Mask Shop entered — switching to MM, slot " << fileNum << std::endl;
-    if (MM_SetComboEntryIsResume)
-        MM_SetComboEntryIsResume(0); // portal entry: always arrives in South Clock Town
-    g_PendingMMFileNum = fileNum;
-    // OOT game loop is already exiting (gGameState->running = false set by the hook).
-}
-
-// Why MM handed control back: 0 = portal, 1 = Ctrl+R reset, 2 = owl-save quit (see BenPort.cpp).
-static int g_mmReturnKind = 0;
-
-static void Combo_OnMMReturn(int kind) {
-    g_mmReturnKind = kind;
-    std::cout << "[ComboShip] MM returning to OOT (kind=" << kind << ")" << std::endl;
-    g_pendingOOTReturn = true;
-}
 
 // ---------- Entry point ----------
