+++ combo/core/ComboSlotBind.h
@@ -1,0 +1,5 @@
+// ComboShip: binding a seed to a save slot - the new-file and load-file callbacks OOT fires.
+#pragma once
+
+void Combo_OnOOTSaveInit(int fileNum);
+void Combo_OnOOTSaveLoad(int fileNum);
